+++ service/src/main/java/eu/europa/ec/fisheries/uvms/plugins/inmarsat/HelperFunctions.java
@@ -1,5 +1,6 @@
 package eu.europa.ec.fisheries.uvms.plugins.inmarsat;
 
+import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.InmarsatSocketException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
