+++ service/src/main/java/eu/europa/ec/fisheries/uvms/plugins/inmarsat/InmarsatPollHandler.java
@@ -4,10 +4,7 @@ import eu.europa.ec.fisheries.schema.exchange.common.v1.AcknowledgeTypeType;
 import eu.europa.ec.fisheries.schema.exchange.common.v1.CommandType;
 import eu.europa.ec.fisheries.schema.exchange.common.v1.KeyValueType;
 import eu.europa.ec.fisheries.schema.exchange.plugin.types.v1.PollType;
-import eu.europa.ec.fisheries.schema.exchange.plugin.types.v1.PollTypeType;
-import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.ConfigPoll;
-import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.InmarsatPoll;
-import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.ManualPoll;
+import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.InmarsatPendingResponse;
 import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.StatusEnum;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,7 +12,6 @@ import org.slf4j.LoggerFactory;
 import javax.ejb.Singleton;
 import javax.inject.Inject;
 import java.io.BufferedInputStream;
-import java.io.InputStream;
 import java.io.PrintStream;
 import java.net.Socket;
 import java.util.ArrayList;
@@ -24,7 +20,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Inmarsat har ikke definert Polling&Data rapport tjenesten hundre prosent, med det mener jeg at terminal leverandører
- * har hatt mulighet til å lage sine «tolkninger» på endel punkter. Hvis du ser på data rapport pakken (PosRep-C_2l.doc)
+ * har hatt mulighet til å lage sine "tolkninger" på endel punkter. Hvis du ser på data rapport pakken (PosRep-C_2l.doc)
  * så ser du at nederst i pakke en er MEM-code. Den mest brukte MEM code er 11 I attribute bytene nederst i pakke en
  * vil mem 11 angi at atribute bytene inneholder  «time of position».
  *
@@ -39,7 +35,7 @@ import java.util.concurrent.ConcurrentHashMap;
  * poll 0,G,4661,N,1,0,4,,5611,24
  * The P8 parameter is the start frame. The start time must be calculated according to the formula:
  * (((hour*60)+minute)*60)/8.64 = start frame number. (night and day (24 hours) are divided in  10000 frame's a'8.64 sec.).
- * The “24” at the end indicate that the terminals shall send one report every hour.
+ * The "24" at the end indicate that the terminals shall send one report every hour.
  *
  * 2)  Afterwards you have to start the reporting command:
  * e.g:  poll 0,G,4661,D,1,0,5
