+++ service/src/main/java/eu/europa/ec/fisheries/uvms/plugins/inmarsat/data/InmarsatPendingResponse.java
@@ -9,10 +9,9 @@ the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the impl
 FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details. You should have received a
 copy of the GNU General Public License along with the IFDM Suite. If not, see <http://www.gnu.org/licenses/>.
  */
-package eu.europa.ec.fisheries.uvms.plugins.inmarsat;
+package eu.europa.ec.fisheries.uvms.plugins.inmarsat.data;
 
 import eu.europa.ec.fisheries.schema.exchange.plugin.types.v1.PollType;
-import eu.europa.ec.fisheries.uvms.plugins.inmarsat.data.StatusEnum;
 
 import java.io.Serializable;
